implement permission guard on remove pet

+++ app/src/module/pet/provider/remove.pet.provider.ts
@@ -1,20 +1,35 @@
 import { NotFoundException } from '@nestjs/common'
 
+import { isNil } from 'lodash'
+
+import { PermissionProvider } from '@/module/permission/provider/permission.provider'
+
 import { RemovePetResponse } from '../pet.response'
-import type { PetRepository } from '../repository/pet.repository'
+import { PetRepository } from '../repository/pet.repository'
 
 export class RemovePetProvider {
-  private readonly empty = 0
+  constructor(
+    private readonly permission: PermissionProvider,
+    private readonly repository: PetRepository
+  ) {}
 
-  constructor(private readonly repository: PetRepository) {}
+  async run(id: string, current: string): Promise<RemovePetResponse> {
+    const pet = await this.repository.find({ _id: id })
 
-  async run(id: string, user: string): Promise<RemovePetResponse> {
-    const amount = await this.repository.remove({ _id: id })
-
-    if (this.empty === amount) {
+    if (isNil(pet)) {
       throw new NotFoundException()
     }
 
-    return new RemovePetResponse(id)
+    const { user, organization } = pet
+
+    const permission = await this.permission.run(current, String(user), String(organization))
+
+    if (permission) {
+      await this.repository.remove({ _id: id })
+
+      return new RemovePetResponse(id)
+    }
+
+    throw new NotFoundException()
   }
 }
